\name{epistasis-internal}
\title{epistasis-internal}

\description{
  Internal epistasis functions. These are not intended to be called by the users.
}

\author{Pariya Behrouzi, Danny Arends, and Ernst C. Wit \cr
Maintainer: Pariya Behrouzi \email{pariya.behrouzi@gmail.com}
}

\keyword{internal}
