\name{epistasis-package}
\alias{epistasis-package}
\docType{package}
\title{
  Detecting Epistatic Selection with Partially Observed Genotype Data
}

\description{
A package for reconstructing an underlying network of genomic signatures of high-dimensional epistatic selection from multi-locus genotype data. 
The package is implemented the recent improvement in the analysis of high-dimensional partially observed genotype data Behrouzi and Wit (2016). 
The network captures the conditionally dependent short- and long-range linkage disequilibrium structure of a genomes and reveals aberrant marker-marker 
associations that are due to epistatic selection rather than gametic linkage.
}

\author{Pariya Behrouzi, Danny Arends, and Ernst C. Wit \cr
Maintainers: Pariya Behrouzi \email{pariya.behrouzi@gmail.com} \cr
}

\references{
P. Behrouzi and E. C. Wit. Detecting Epistatic Selection with Partially Observed Genotype Data Using Copula Graphical Models. \emph{arXiv}, 2016.\cr}

\keyword{package}