\name{print.epi}
\alias{print.epi}
\title{
Print function for S3 class "epi"
}
\description{
Print a summary of simulated data from function \code{\link{epistasis}}. 
}
\usage{
\method{print}{epi}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"epi"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}

\author{ 
  Pariya Behrouzi and Ernst C. Wit \cr
  Maintainer: Pariya Behrouzi \email{pariya.behrouzi@gmail.com} \cr
}

\references{
P. Behrouzi and E. C. Wit. Detecting Epistatic Selection with Partially Observed Genotype Data Using Copula Graphical Models. \emph{arXiv}, 2016.\cr
}

\seealso{
\code{\link{epistasis}} 
}