\name{print.episim}
\alias{print.episim}
\title{
Print function for S3 class "episim"
}
\description{
Print a summary of simulated data from function \code{\link{episim}}. 
}
\usage{
\method{print}{episim}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"episim"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}

\author{ 
  Pariya Behrouzi and Ernst C. Wit \cr
  Maintainer: Pariya Behrouzi \email{pariya.behrouzi@gmail.com} \cr
}

\references{
P. Behrouzi and E. C. Wit. Detecting Epistatic Selection with Partially Observed Genotype Data Using Copula Graphical Models. \emph{arXiv}, 2016.\cr
}

\seealso{
\code{\link{episim}} 
}